Sort list of projects before delivering to appsource (#2041)
### ❔What, Why & How

<!-- Include description of the changes that will help reviewers in
their task -->
Currently the Deliver function iterates over the AL-Go projects in the
order they are returned by GetProjectsFromRepository. However, the
output of this function is not sorted by dependencies. E.g. in BCApps
the order would be:

- build\projects\Apps (W1)
- build\projects\Business Foundation Tests
- build\projects\Performance Toolkit Tests
- build\projects\System Application
- build\projects\System Application Modules
- build\projects\System Application Tests
- build\projects\Test Stability Tools

This can be an issue when Publishing to AppSource if there are
dependencies between the apps you are deploying.

Proposed solution: Deliver in the same order as the build. That would
change the order to:

- build\projects\System Application
- build\projects\Test Stability Tools
- build\projects\Apps (W1)
- build\projects\Business Foundation Tests
- build\projects\Performance Toolkit Tests
- build\projects\System Application Modules
- build\projects\System Application Tests

Related to issue: #2004

### ✅ Checklist

- [ ] Add tests (E2E, unit tests)
- [x] Update RELEASENOTES.md
- [ ] Update documentation (e.g. for new settings or scenarios)
- [ ] Add telemetry

<!-- Include more checklist entries, if needed -->

---------

Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>