[Suggestion] Add the ability to compile all test apps in one job and run the tests in a following job  (#2066)
### ❔What, Why & How

<!-- Include description of the changes that will help reviewers in
their task -->
We're trying to see if we can adjust our build graph such that we can
compile everything in the first build matrix and then run all test tests
in the following build matrix. However, currently it seems that AL-Go
won't run tests for test apps downloaded as dependencies (from another
job). This PR would add the ability to change that (as a preview
feature). This may be something we'd change later on though

Related to
[AB#613617](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/613617)

### ✅ Checklist

- [ ] Add tests (E2E, unit tests)
- [x] Update RELEASENOTES.md
- [x] Update documentation (e.g. for new settings or scenarios)
- [ ] Add telemetry

<!-- Include more checklist entries, if needed -->